feat: Allow filtering subnets by any tag
This change allows filtering the list of subnets within which EC2
instances are launched.

Previously, they could only be filtered by name, and only using prefix
matching, something like "my-subnet-*".

The second restriction is still in place, however now any tag can be
used, which means arbitrary collections of subnets can be selected as
long as a tag enumerating them exists on the subnets.

This requires a new configuration variable, which defaults to the subnet
`Name` tag to ensure backwards compatibility.